backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openai-whisper
python-multipart
sounddevice
numpy
scipy
ffmpeg-python 
google-generativeai>=0.5.0
soundfile